requirements.txt
langchain_community
streamlit
pandas
langchain
huggingface_hub
python-dotenv